feat(resolver): Move all predicates to its own file
