+++ pkg/controller/registry/resolver/cache.go
@@ -8,7 +8,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/blang/semver/v4"
 	"github.com/sirupsen/logrus"
 
 	"github.com/operator-framework/operator-lifecycle-manager/pkg/api/client/listers/operators/v1alpha1"
@@ -387,16 +386,6 @@ func (s SortableSnapshots) Swap(i, j int) {
 	s.snapshots[i], s.snapshots[j] = s.snapshots[j], s.snapshots[i]
 }
 
-type OperatorPredicateFunc func(*Operator) bool
-
-func (opf OperatorPredicateFunc) Test(o *Operator) bool {
-	return opf(o)
-}
-
-type OperatorPredicate interface {
-	Test(*Operator) bool
-}
-
 func (s *CatalogSnapshot) Find(p ...OperatorPredicate) []*Operator {
 	s.m.RLock()
 	defer s.m.RUnlock()
@@ -418,155 +407,6 @@ type EmptyOperatorFinder struct{}
 
 func (f EmptyOperatorFinder) Find(...OperatorPredicate) []*Operator {
 	return nil
-}
-
-func WithCSVName(name string) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		return o.name == name
-	})
-}
-
-func WithChannel(channel string) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		// all operators match the empty channel
-		if channel == "" {
-			return true
-		}
-		if o.bundle == nil {
-			return false
-		}
-		return o.bundle.ChannelName == channel
-	})
-}
-
-func WithPackage(pkg string) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		for _, p := range o.Properties() {
-			if p.Type != opregistry.PackageType {
-				continue
-			}
-			var prop opregistry.PackageProperty
-			err := json.Unmarshal([]byte(p.Value), &prop)
-			if err != nil {
-				continue
-			}
-			if prop.PackageName == pkg {
-				return true
-			}
-		}
-		return o.Package() == pkg
-	})
-}
-
-func WithVersionInRange(r semver.Range) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		for _, p := range o.Properties() {
-			if p.Type != opregistry.PackageType {
-				continue
-			}
-			var prop opregistry.PackageProperty
-			err := json.Unmarshal([]byte(p.Value), &prop)
-			if err != nil {
-				continue
-			}
-			ver, err := semver.Parse(prop.Version)
-			if err != nil {
-				continue
-			}
-			if r(ver) {
-				return true
-			}
-		}
-		return o.version != nil && r(*o.version)
-	})
-}
-
-func WithLabel(label string) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		for _, p := range o.Properties() {
-			if p.Type != opregistry.LabelType {
-				continue
-			}
-			var prop opregistry.LabelProperty
-			err := json.Unmarshal([]byte(p.Value), &prop)
-			if err != nil {
-				continue
-			}
-			if prop.Label == label {
-				return true
-			}
-		}
-		return false
-	})
-}
-
-func WithCatalog(key registry.CatalogKey) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		return key.Equal(o.SourceInfo().Catalog)
-	})
-}
-
-func ProvidingAPI(api opregistry.APIKey) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		for _, p := range o.Properties() {
-			if p.Type != opregistry.GVKType {
-				continue
-			}
-			var prop opregistry.GVKProperty
-			err := json.Unmarshal([]byte(p.Value), &prop)
-			if err != nil {
-				continue
-			}
-			if prop.Kind == api.Kind && prop.Version == api.Version && prop.Group == api.Group {
-				return true
-			}
-		}
-		return false
-	})
-}
-
-func SkipRangeIncludes(version semver.Version) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		// TODO: lift range parsing to OperatorSurface
-		semverRange, err := semver.ParseRange(o.bundle.SkipRange)
-		return err == nil && semverRange(version)
-	})
-}
-
-func Replaces(name string) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		if o.Replaces() == name {
-			return true
-		}
-		for _, s := range o.bundle.Skips {
-			if s == name {
-				return true
-			}
-		}
-		return false
-	})
-}
-
-func And(p ...OperatorPredicate) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		for _, l := range p {
-			if l.Test(o) == false {
-				return false
-			}
-		}
-		return true
-	})
-}
-
-func Or(p ...OperatorPredicate) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		for _, l := range p {
-			if l.Test(o) == true {
-				return true
-			}
-		}
-		return false
-	})
 }
 
 func AtLeast(n int, operators []*Operator) ([]*Operator, error) {
@@ -595,26 +435,4 @@ func Filter(operators []*Operator, p ...OperatorPredicate) []*Operator {
 
 func Matches(o *Operator, p ...OperatorPredicate) bool {
 	return And(p...).Test(o)
-}
-
-func True() OperatorPredicate {
-	return OperatorPredicateFunc(func(*Operator) bool {
-		return true
-	})
-}
-
-func False() OperatorPredicate {
-	return OperatorPredicateFunc(func(*Operator) bool {
-		return false
-	})
-}
-
-func CountingPredicate(p OperatorPredicate, n *int) OperatorPredicate {
-	return OperatorPredicateFunc(func(o *Operator) bool {
-		if p.Test(o) {
-			*n++
-			return true
-		}
-		return false
-	})
 }
