+++ pkg/controller/registry/resolver/cache_test.go
@@ -339,49 +339,3 @@ func TestStripPluralRequiredAndProvidedAPIKeys(t *testing.T) {
 	assert.Equal(t, "K.v1.g", result[0].providedAPIs.String())
 	assert.Equal(t, "K2.v2.g2", result[0].requiredAPIs.String())
 }
-
-func TestCountingPredicate(t *testing.T) {
-	for _, tc := range []struct {
-		Name        string
-		TestResults []bool
-		Expected    int
-	}{
-		{
-			Name:        "no increment on failure",
-			TestResults: []bool{false},
-			Expected:    0,
-		},
-		{
-			Name:        "increment on success",
-			TestResults: []bool{true},
-			Expected:    1,
-		},
-		{
-			Name:        "multiple increments",
-			TestResults: []bool{true, true},
-			Expected:    2,
-		},
-		{
-			Name:        "no increment without test",
-			TestResults: nil,
-			Expected:    0,
-		},
-	} {
-		t.Run(tc.Name, func(t *testing.T) {
-			var (
-				n      int
-				result bool
-			)
-
-			p := CountingPredicate(OperatorPredicateFunc(func(*Operator) bool {
-				return result
-			}), &n)
-
-			for _, result = range tc.TestResults {
-				p.Test(nil)
-			}
-
-			assert.Equal(t, tc.Expected, n)
-		})
-	}
-}
